Prepare version  0.12.0: * move IconRegistry to api * increase version to 0.12.0 * remove limitation regarding single column properties
