+++ src/main/kotlin/com/github/chriskn/structurizrextension/api/model/C4Properties.kt
@@ -1,6 +1,6 @@
 package com.github.chriskn.structurizrextension.api.model
 
-private const val MAX_ROW_SIZE = 4
+private const val MAX_COLUMN_SIZE = 4
 
 /**
  * Properties can be defined for Elements and Dependencies.
@@ -8,7 +8,7 @@ private const val MAX_ROW_SIZE = 4
  * They will be rendered as table and can be used to document detailed concepts like deployments.
  * By default, no (table) header is added (WithoutPropertyHeader() in C4-PlantUML).
  * The header must have at least the number of columns of the longest row.
- * Number rows should not exceed [MAX_ROW_SIZE] or number of header rows
+ * Number columns should not exceed [MAX_COLUMN_SIZE] or number of header rows
  * C4PlantUml does not support single row properties.
  */
 data class C4Properties(
@@ -23,12 +23,8 @@ data class C4Properties(
 ) {
 
     init {
-        require(values.all { it.size <= MAX_ROW_SIZE && (header == null || it.size <= header.size) }) {
-            "Number of rows should not exceed $MAX_ROW_SIZE or number of header rows"
-        }
-        // Not supported. Issue: https://github.com/plantuml-stdlib/C4-PlantUML/issues/355
-        require(values.all { it.size > 1 } && (header == null || header.size > 1)) {
-            "C4PlantUml does not support single row properties or property header"
+        require(values.all { it.size <= MAX_COLUMN_SIZE && (header == null || it.size <= header.size) }) {
+            "Number of columns should not exceed $MAX_COLUMN_SIZE or number of header rows"
         }
     }
 }
